+++ src/Metamorph/MacroHandler.php
@@ -12,9 +12,14 @@ namespace DecodeLabs\Metamorph;
 interface MacroHandler extends Handler
 {
     /**
+     * @var array<string,array<string,mixed>>
+     */
+    public const array Macros = [];
+
+    /**
      * Load macro options
      *
-     * @return array<string, mixed>|null
+     * @return array<string,mixed>|null
      */
     public static function loadMacro(
         string $name